wokkahlearn/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
django-filter==23.3
django-redis==5.4.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Celery
celery==5.3.1
django-celery-beat==2.5.0
kombu==5.3.2

# File Storage
django-storages==1.14.2
boto3==1.28.57

# Image Processing
Pillow==10.0.1

# Code Execution
docker==6.1.3

# AI/ML
openai==0.28.1
anthropic==0.3.11
scikit-learn==1.3.0
pandas==2.1.1
numpy==1.25.2

# Web Server
gunicorn==21.2.0
gevent==23.7.0

# Monitoring
sentry-sdk==1.32.0
prometheus-client==0.17.1

# Development
python-decouple==3.8
whitenoise==6.5.0

